refactor : seperate CSR and SSR of article edit page
