+++ src/app/article/edit/[id]/page.tsx
@@ -1,18 +1,13 @@
-'use client'
+import { notFound } from 'next/navigation'
 
-import { notFound, useRouter } from 'next/navigation'
+import { getArticleById, GetArticleResponseInterface } from '@/apis/articles'
+import ArticleEdit from '@/containers/Article/Edit'
 
-import {
-  ArticleInterface,
-  getArticleById,
-  GetArticleResponseInterface,
-  putArticleById,
-} from '@/apis/articles'
-
-import ArticleForm from '@/components/ArticleForm'
-
-const EditArticle = async ({ params: { id } }: { params: { id: string } }) => {
-  const router = useRouter()
+const ArticleEditPage = async ({
+  params: { id },
+}: {
+  params: { id: string }
+}) => {
   const loadedArticle = async (): Promise<
     GetArticleResponseInterface | undefined
   > => {
@@ -31,37 +26,8 @@ const EditArticle = async ({ params: { id } }: { params: { id: string } }) => {
 
   const data = await loadedArticle()
   if (!data) return notFound()
-  const {
-    article: { title: originalTitle, content: originalContent },
-  } = data
+  const { article } = data
 
-  const handleSubmit = async (article: ArticleInterface) => {
-    const { title: newTitle, content: newContent } = article
-
-    if (!newTitle || !newContent) {
-      alert('Title and description are required')
-      return
-    }
-
-    try {
-      const res = await putArticleById(id, { newTitle, newContent })
-
-      if (!res.ok) {
-        throw new Error(`HTTP error! Status: ${res.status}`)
-      }
-
-      router.push(`/${id}`)
-    } catch (error) {
-      console.log(error)
-    }
-  }
-
-  return (
-    <ArticleForm
-      title={originalTitle}
-      content={originalContent}
-      onSubmit={handleSubmit}
-    />
-  )
+  return <ArticleEdit article={article} />
 }
-export default EditArticle
+export default ArticleEditPage
